Do not block the main thread waiting for images.
If the page that is to be opened has not been extracted yet, simply
display an empty pixbuf in its stead, and wait for the callback
to signal that the page has been extracted. This should allow for
faster startup, and might also enable extractor threads to post
events to the main thread, e.g. for displaying a password prompt dialog.

git-svn-id: https://mcomix.svn.sourceforge.net/svnroot/mcomix@265 6cf541bf-b722-4a7c-9fc0-fb1e75421f59