Confirmation dialog cleanup.
Removed several confirmation dialogs for non-destructive operations,
such as removing files from the library or deleting collections. Style
for other message dialogs has been unified, and "Yes" has been set as
default answer for most dialogs.

This goes in line with the thought that the user generally isn't stupid
and really wants to perform an action when he goes to the trouble of
clicking on it.

git-svn-id: https://mcomix.svn.sourceforge.net/svnroot/mcomix@434 6cf541bf-b722-4a7c-9fc0-fb1e75421f59